+++ src/web3mail/fetchMyContacts.ts
@@ -1,9 +1,10 @@
-import { throwIfMissing } from '../utils/validators.js';
+import { booleanSchema, throwIfMissing } from '../utils/validators.js';
 import { fetchUserContacts } from './fetchUserContacts.js';
 import {
   Contact,
   DappAddressConsumer,
   DappWhitelistAddressConsumer,
+  FetchMyContactsParams,
   IExecConsumer,
   SubgraphConsumer,
 } from './types.js';
@@ -13,10 +14,16 @@ export const fetchMyContacts = async ({
   iexec = throwIfMissing(),
   dappAddressOrENS = throwIfMissing(),
   dappWhitelistAddress = throwIfMissing(),
+  isUserStrict = false,
 }: IExecConsumer &
   SubgraphConsumer &
   DappAddressConsumer &
-  DappWhitelistAddressConsumer): Promise<Contact[]> => {
+  DappWhitelistAddressConsumer &
+  FetchMyContactsParams): Promise<Contact[]> => {
+  const vIsUserStrict = booleanSchema()
+    .label('isUserStrict')
+    .validateSync(isUserStrict);
+
   const userAddress = await iexec.wallet.getAddress();
   return fetchUserContacts({
     iexec,
@@ -24,5 +31,6 @@ export const fetchMyContacts = async ({
     dappAddressOrENS,
     dappWhitelistAddress,
     userAddress,
+    isUserStrict: vIsUserStrict,
   });
 };